Implement delete functionality for event organiser
This is the basic delete function, to be integrated with UI.

+++ src/main/java/uk/ac/man/cs/eventlite/controllers/EventsController.java
@@ -71,7 +71,7 @@ public class EventsController {
 	
 	@PostMapping("/deleted")
 	public String deleteEvent(HttpServletRequest request) {
-		eventService.deleteById(Integer.parseInt(request.getParameter("eventID")));		
+		eventService.deleteById(Long.parseLong(request.getParameter("eventID")));		
 		return "redirect:/events";
