+++ src/main/java/uk/ac/man/cs/eventlite/dao/EventRepository.java
@@ -11,5 +11,5 @@ public interface EventRepository extends CrudRepository<Event, Long> {
 	
 	public Iterable<Event> findAllByOrderByDateAscTimeAsc();
 	
-	public void deleteById(int id);
+	public void deleteById(long id);
 }
