+++ src/main/java/uk/ac/man/cs/eventlite/dao/EventService.java
@@ -10,5 +10,5 @@ public interface EventService {
 	
 	public Event save(Event e);
 	
-	public void deleteById(int id);
+	public void deleteById(long id);
 }
